Overhaul of SDK and exts
READ THIS COMMIT DESCRIPTION if you encounter build errors after
upgrading!
This is a breaking update which introduces some very useful features.

• Added C++11 type checking at runtime.
C++ function return types and parameter types will be checked against
the corresponding JSON on Run Application. This is done in DarkEDIF.h
with vardiac templates, odd use of static types, and recursive template
loops – only in Debug mode, though.
This check is automatically enabled in Debug or Debug Unicode builds by
redefine of LinkXXX preprocessor macros. You can see their magnificence
in DarkEDIF.h.
In all other build variants, the regular LinkXXX will be used and no
check will be done.

• Revamp of main header to use enum classes. Enum classes are newer C++
and require VS 2015 or above.
• Fixed Combo Box property. Stores index in edPtr instead of full
string. Fusion supplies index, so this is overdue.
• Added the ability to use Unsigned Integer in JSON. Fusion won't read
the type as unsigned, but for Fusion 3 preparation, types can be left as
genuine as possible.
• Added three useful functions for getting a property's value in
EDITDATA. Some modification to Common.h is needed. Only usable during
Extension ctor.

To upgrade your DarkEDIF extensions:
1) You need VS 2015 or greater.
2) Convert your Handle classes to return REFLAG; there's a REFLAG::NONE
value for 0.
3) Change the variable types of OEPREFS and OEFLAGS in Extension.h; it
should read "OEFLAGS OEFLAGS = OEFLAGS::NONE;" or what have you.
4) Copy the four lines on the bottom of EDITDATA (with the three
property-reading functions and associated comment) into your EDITDATA
struct.

• Corrected variable naming in Lacewing Blue Server and Client. Now
they're lowerCamelCase instead of PascalCase.
• Removed unnecessary null pointer checks in A/C/E – DarkEDIF checks
text parameter values in Edif.cpp. They can still be blank (first char
null), obviously, but the pointers themselves can't be null.
• Corrected the bizarre spacing of the Lacewing JSONs. They're a bit
easier to look at now.
• Fixed typos in Bluewing Server French translation. Not currently
published to CEM/Darkwire.

• Added Klystron pattern/music player extension. It doesn't load any
songs, but builds OK. The loading problem is due to a fault of the base
library, which may be fixed later.
• Fixed rSDK TabControl to use correct project name, and downgraded VS
project version to match rest of rSDK. No newer VS features were used
anyway.

+++ DarkEDIF/Lib/Lacewing/RelayClient.cc
@@ -1284,7 +1284,7 @@ namespace lacewing
 		return (p.begin() == p.end()) ? nullptr : &(*p.begin())->public_;
 	}
 
-	int relayclient::channellistingcount() const
+	size_t relayclient::channellistingcount() const
 	{
 		return ((relayclientinternal *)internaltag)->channellist.size();
 	}
